+++ main/main.go
@@ -2,7 +2,7 @@ package main
 
 import (
 	"fmt"
-	//display "probsolbygolang/displayTrees"
+	//"probsolbygolang/displayTrees"
 	//"probsolbygolang/problems"
 )
 
@@ -16,11 +16,17 @@ func main() {
 	root := problems.SortedArrayToBST(arr, 0, n-1)
 
 	fmt.Println("PreOrder Traversal of the balanced BST: (Root, Left, Right)")
-	display.PreOrder(root)
+	displayTrees.PreOrder(root)
 	fmt.Println("\nInOrder Traversal of the balanced BST: (Left, Root, Right)")
-	display.InOrder(root)
+	displayTrees.InOrder(root)
 	fmt.Println("\nPostOrder Traversal of the balanced BST: (Left, Right, Root)")
-	display.PostOrder(root)*/
+	displayTrees.PostOrder(root)
 
+	fmt.Println("")
+	arr = []int{6, 4, 12, 10, 22, 54, 32, 42, 21, 11}
+	pairsWithEqualSum := problems.PairsWithEqualSum(arr)
+	for _, val := range pairsWithEqualSum {
+		fmt.Println(val)
+	}*/
 	fmt.Println("\nThank you...")
 }