+++ CHANGELOG.md
@@ -3,3 +3,9 @@
 ## Added (5th July 2021)
 - Convert a sorted integer array to balanced Binary Search Tree
 - Preorder display of BST
+- Postorder display of BST
+- Inorder display of BST
+- Print the pairs of elements with equal sum in a given array
+
+## Fixed (5th July 2021)
+- Package name issues in model/BST_Node.go